[improvement] bottom navigation cleaned
